Bug 1504452 - fix GrGLGpu buffer offset calculations. r=rhunt

+++ gfx/skia/skia/src/gpu/gl/GrGLGpu.cpp
@@ -1823,7 +1823,7 @@ void GrGLGpu::setupGeometry(const GrBuffer* indexBuffer,
 
     if (int vertexStride = fHWProgram->vertexStride()) {
         SkASSERT(vertexBuffer && !vertexBuffer->isMapped());
-        size_t bufferOffset = vertexBuffer->baseOffset() + baseVertex * vertexStride;
+        size_t bufferOffset = vertexBuffer->baseOffset() + baseVertex * size_t(vertexStride);
         for (int i = 0; i < fHWProgram->numVertexAttributes(); ++i) {
             const auto& attrib = fHWProgram->vertexAttribute(i);
             static constexpr int kDivisor = 0;
@@ -1833,7 +1833,7 @@ void GrGLGpu::setupGeometry(const GrBuffer* indexBuffer,
     }
     if (int instanceStride = fHWProgram->instanceStride()) {
         SkASSERT(instanceBuffer && !instanceBuffer->isMapped());
-        size_t bufferOffset = instanceBuffer->baseOffset() + baseInstance * instanceStride;
+        size_t bufferOffset = instanceBuffer->baseOffset() + baseInstance * size_t(instanceStride);
         int attribIdx = fHWProgram->numVertexAttributes();
         for (int i = 0; i < fHWProgram->numInstanceAttributes(); ++i, ++attribIdx) {
             const auto& attrib = fHWProgram->instanceAttribute(i);
